+ PUT /tereni/{id}
request body requires an object like:
{
    "naziv":"teren dva",
    "vlasnikTerenaId":1,
    "tip":"vanjski"
}

+++ IzvorniKod/ppadel-spring/src/main/java/fer/progi/mjesecari/ppadel/api/TerenController.java
@@ -1,6 +1,7 @@
 package fer.progi.mjesecari.ppadel.api;
 
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import fer.progi.mjesecari.ppadel.dao.TerenRepository;
 import fer.progi.mjesecari.ppadel.domain.Teren;
@@ -13,8 +14,12 @@ import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 
+import javax.print.attribute.standard.PrinterMessageFromOperator;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Profile;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PostAuthorize;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
@@ -22,6 +27,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
@@ -56,6 +62,10 @@ public class TerenController {
     @PostAuthorize("hasRole('ROLE_OWNER')")
     @DeleteMapping("/{id}")
     public void removeTeren(Principal principal, @PathVariable Long id) {
+        String mail = mailFromPrincipal(principal);
+        if(mail == null || !terenService.fetch(id).getVlasnikTeren().getEmail().equals(mail)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         terenService.deleteTeren(id);
     }
 
@@ -67,16 +77,38 @@ public class TerenController {
     @PostAuthorize("hasRole('ROLE_OWNER')")
     @GetMapping("/my")
     public Collection<Teren> myTereni(Principal principal) {
-        if( OAuth2AuthenticationToken.class.isInstance(principal) ){
-            Map<String, Object> attributes = ((OAuth2AuthenticationToken) principal).getPrincipal().getAttributes();
-            String email = (String) attributes.get("email");
-
-            return terenRepo.findAllByVlasnikTerenEmail(email);
-
-        }
-        return new ArrayList<Teren>();
+        String mail = mailFromPrincipal(principal);
+        if(mail != null)
+            return terenRepo.findAllByVlasnikTerenEmail(mail);
+        else
+            return new ArrayList<Teren>();
     }
     
+    @PostAuthorize("hasRole('ROLE_OWNER')")
+    @PutMapping("/{id}")
+    public Teren editTeren(Principal principal, @PathVariable Long id, @RequestBody createTerenDTO terenDto) {
+        String mail = mailFromPrincipal(principal);
+        
+        if(mail == null || !terenService.fetch(id).getVlasnikTeren().getEmail().equals(mail)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+        else{
+            terenService.updateTerenName(id, terenDto.getNaziv());
+            return terenService.updateTerenType(id, terenDto.getTip());
+        }
+
+    }
+    
+
+    private String mailFromPrincipal(Principal principal){
+        if( OAuth2AuthenticationToken.class.isInstance(principal) ){
+            Map<String, Object> attributes = ((OAuth2AuthenticationToken) principal).getPrincipal().getAttributes();
+            return (String) attributes.get("email");
+        }
+        else {
+            return null;
+        }
+    }
